Correção do erro de carregamento
