+++ semana12/labeddit/src/screens/FeedPage/FeedPage.js
@@ -1,12 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios'
-import { Container, makeStyles, Typography } from '@material-ui/core';
+import { Container, makeStyles, Typography, Button, CircularProgress, TextField } from '@material-ui/core';
 import useProtectedPage from '../../hooks/useProtectedPage';
 import PostCard from '../../components/PostCard/PostCard';
 import Loading from '../../components/Loading/Loading';
 import { timePassed } from '../../helpers/timePassed'
 import { baseURL } from '../../constants/urls';
-import FeedForm from './FeedForm';
+import useForm from '../../hooks/useForm'
+import { createPost } from '../../services/posts'
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -21,6 +22,8 @@ const useStyles = makeStyles((theme) => ({
 const FeedPage = () => {
     const [postArray, setPostArray] = useState([])
     const classes = useStyles();
+    const [form, handleInputChange, resetState] = useForm({ title: '', text: ''})
+    const [isLoading, setIsLoading] = useState(false)
     
     useProtectedPage();
     useEffect(() => {
@@ -30,21 +33,23 @@ const FeedPage = () => {
     const getAllPosts = () => {
         axios.get(`${baseURL}/posts`, {
             headers: {
-              Authorization: localStorage.getItem('token')
+                Authorization: localStorage.getItem('token')
             }
-          })
-            .then((response) => {
-              setPostArray(response.data.posts)
-            })
-            .catch((error) => {
-              console.log(error)
-              alert('Ocorreu um erro, tente novamente')
-            })
+        })
+        .then((response) => {
+            setPostArray(response.data.posts)
+        })
+        .catch((error) => {
+            console.log(error)
+            alert('Ocorreu um erro, tente novamente')
+        })
     }
-    
-    const renderPosts = () => {
+        
+        const renderPosts = () => {
         return (
-            postArray.sort((a, b) => {return b.createdAt - a.createdAt}).map((post) => {
+            postArray.filter(post => {return typeof post.title === 'string'} )
+            .sort((a, b) => {return b.createdAt - a.createdAt})
+            .map(post => {
                 return (
                     <PostCard 
                         key={post.id} 
@@ -61,13 +66,54 @@ const FeedPage = () => {
         )
     }
 
+    const onClickCreate = (event) => {
+        event.preventDefault()
+        const element = document.getElementById('feed-form')
+        const isValid = element.checkValidity()
+        element.reportValidity()
+        if(isValid){
+            createPost(form, setIsLoading, getAllPosts)
+            resetState()
+            
+        }
+    }
+
     return ( 
         <Container className={classes.root} maxWidth="md">
             <Typography variant={'h5'} align={"center"}>Crie seu post</Typography>
             <Container className={classes.subcontainer}>
-                <FeedForm getPosts={getAllPosts}/>
+                <form id='feed-form'>
+                    <TextField
+                    value={form.title}
+                    onChange={handleInputChange}
+                    margin="normal"
+                    required
+                    fullWidth
+                    id="title"
+                    label="Título do post"
+                    name="title"
+                    />
+                    <TextField
+                    value={form.text}
+                    onChange={handleInputChange}
+                    margin="normal"
+                    required
+                    fullWidth
+                    id="text"
+                    label="Escreva aqui seu post"
+                    name="text"
+                    />
+                    <Button
+                        type="submit"
+                        variant="contained"
+                        color="primary"
+                        onClick={onClickCreate}
+                    >
+                        {isLoading ? <CircularProgress color={'inherit'} size={24}/> : <>postar</>}
+                    </Button>
+                </form>
             </Container>
-            {postArray.length > 0 ? renderPosts() : <Loading/>}
+            {postArray && postArray.length > 0 ? renderPosts() : <Loading/>}
         </Container>
      );
 }
