+++ semana12/labeddit/src/screens/PostPage/PostPage.js
@@ -1,13 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { Container, makeStyles } from '@material-ui/core';
+import { Container, makeStyles, Button, CircularProgress, TextField } from '@material-ui/core';
 import PostCard from '../../components/PostCard/PostCard';
 import CommentBox from '../../components/CommentBox'
 import Loading from '../../components/Loading/Loading';
 import useProtectedPage from '../../hooks/useProtectedPage';
 import { getPostDetail } from '../../services/posts'
 import { timePassed } from '../../helpers/timePassed'
-import CommentForm from './CommentForm'
+import useForm from '../../hooks/useForm';
+import { CommentFormBox } from './styled';
+import { createComment } from '../../services/posts';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -25,15 +27,27 @@ const PostPage = () => {
     const {id} = useParams();
     const [postDetail, setPostDetail] = useState({})
     const [postComments, setPostComments] = useState([])
+    const [form, handleInputChange, resetState] = useForm({ text: ''})
+    const [isLoading, setIsLoading] = useState(false)
 
     useProtectedPage()
     
     useEffect(() => {
         getPostDetail(id, setPostDetail, setPostComments)
     }, [id])
+
+    const onClickCreate = (event) => {
+        event.preventDefault()
+        const element = document.getElementById('comment-form')
+        const isValid = element.checkValidity()
+        element.reportValidity()
+        if(isValid){
+            createComment(id, form, setIsLoading, getPostDetail, setPostDetail, setPostComments)
+            resetState()
+        }
+    }
     
     const renderPostDetail = () => {
-
         return (
             <Container className={classes.root}>
                 <PostCard
@@ -46,9 +60,26 @@ const PostPage = () => {
                     postId={postDetail.id}
                     className={classes.box}
                     />
-                <CommentForm 
-                    postId={id}
-                />
+                <CommentFormBox id='comment-form'>
+                    <TextField
+                        value={form.text}
+                        onChange={handleInputChange}
+                        margin="normal"
+                        required
+                        fullWidth
+                        id="text"
+                        label="Escreva aqui seu comentário"
+                        name="text"
+                    />
+                    <Button
+                        type="submit"
+                        variant="contained"
+                        color="primary"
+                        onClick={onClickCreate}
+                    >
+                        {isLoading ? <CircularProgress color={'inherit'} size={24}/> : <>comentar</>}
+                    </Button>
+                </CommentFormBox>
                     {postComments.sort((a, b) => {return b.createdAt - a.createdAt}).map((comment) => {
                         return (
                             <CommentBox
